Add a context menu to files in a Changeset
- on double click, focus the Content Browser on the selected file
- display the history of the file
- diff selected revision against the previous one
- diff selected revision against the file in the workspace
- menu: explicit entry to browse to the file (content browser) like double click
- save this revision as...
- revert to this revision

Limitations:
- doesn't work as well for older versions of Unreal than 5.2 (cannot find Actors in world partition level etc)
- cm history doesn't work for assets deleted

+++ Source/PlasticSourceControl/Private/SPlasticSourceControlChangesetFileRow.cpp
@@ -5,6 +5,7 @@
 #include "PlasticSourceControlState.h"
 #include "PlasticSourceControlUtils.h"
 
+#include "Misc/Paths.h"
 #include "Widgets/Images/SImage.h"
 #include "Widgets/Images/SLayeredImage.h"
 #include "Widgets/Text/STextBlock.h"
@@ -50,7 +51,7 @@ TSharedRef<SWidget> GetSCCFileWidget(FPlasticSourceControlState* InFileState)
 
 	// Make icon overlays (eg, SCC and dirty status) a reasonable size in relation to the icon size (note: it is assumed this icon is square)
 	const float ICON_SCALING_FACTOR = 0.7f;
-	const float IconOverlaySize = IconBrush->ImageSize.X * ICON_SCALING_FACTOR;
+	const float IconOverlaySize = static_cast<float>(IconBrush->ImageSize.X * ICON_SCALING_FACTOR);
 
 	return SNew(SOverlay)
 		// The actual icon
